Change a few things w.r.t. settings / camera parameters: - implement new generation of initial settings file based on available cameras   Basler camera detection only for now - remove white balance (it is for RGB images, not grayscale images) - frames per second is now typed float - remove "shutter" parameter (flycapture2 calls exposure "shutter", its "auto_exposure" parameter   is the target value for auto exposure/gain mode, which we don't currently model)

+++ src/VideoStream.cpp
@@ -4,7 +4,7 @@
 
 VideoStream::VideoStream(std::string                                  _id,
                          std::tuple<size_t, size_t>                   _resolution,
-                         size_t                                       _framesPerSecond,
+                         float                                        _framesPerSecond,
                          size_t                                       _framesPerFile,
                          std::unordered_map<std::string, std::string> _encoderOptions)
 : id{_id}
